+++ NEWS.md
@@ -1,3 +1,8 @@
+# inbodb 0.0.10
+
+* bugfix: solve encoding problem with database Watina by deprecating argument
+  `autoconvert_utf8` in function `connect_inbo_dbase()` (issue #70)
+
 # inbodb 0.0.9
 
 * bugfix: error on load